+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Test for the matrix client creation.
+- Ability to set the matrix client login method.
+- Matrix room state migrator service.
+
+### Changed
+- Service interface refactored to the CLI interface.
 
 ## [0.1.3] - 2024-06-25
 ### Changed